Final changes to the new logging capabilities.  All print statements removed.
 On branch master
 Your branch is up-to-date with 'origin/master'.

 Changes to be committed:
	modified:   async_re.py
	modified:   bedam_async_re.py
	modified:   bedamtempt_async_re.py
	modified:   boinc_transport.py
	modified:   impact_async_re.py
	modified:   ssh_transport.py
	modified:   tempt_async_re.py
	modified:   transport.py
	new file:   utils/logging.conf

Removed print statements and replaced them with functionality from the python
logging module.  There is a logger.conf file in utils/ which stores the logger
configuration.  The benefits of this is that the logged messages now have
timestamps, meaningful logging levels (warning, critical, info, debug, etc.),
and display from what module they came from.

The logger is set up in such a way that it resembles print statements as much as
possible, so currently it logs everything to stdout.  In the future, this could
be changed to a dedicated log file if so desired.

All the changes were tested in a boinc environment, but not an ssh or multiarch
environment, so there may be some wacky functionality there.  Nothing should be
broken as almost changes were swapping out print statements for pseudo print
statements.